requirements.txt
load-dotenv
python-dotenv
openai
pydantic
fastapi
uvicorn
pyyaml
jinja2